rename FAQPage to AboutPage
