+++ assets/tests/components/aboutPage.test.tsx
@@ -1,0 +1,11 @@
+import React from "react"
+import renderer from "react-test-renderer"
+import AboutPage from "../../src/components/aboutPage"
+
+describe("About page", () => {
+  test("renders", () => {
+    const tree = renderer.create(<AboutPage />).toJSON()
+
+    expect(tree).toMatchSnapshot()
+  })
+})
